Static site generation from XML
src/site.xml has the catalog, and it pulls in the individual pages.
We use XSLT to generate index.html from site.xml, and the model HTML
pages get generated on the side as site.xml is processed.

A rudimentary schema describing the XML format is in
src/schema/site.xsd.

In addition to generating the site HTML, build.py is augmented to copy
assets from both the local repo and a designated assets repo, which
contains the models.

We currently use Saxon for XSLT 3.0 processing.