Promisify ModelController.loadModel
This allows us to attach post-load functionality to the asynchronous
loadModel operation in an easier way than hacking in more
callbacks. We'll use this to process hitbox links after we've loaded
the file.

We can immediately use the Promise interface to hoist the console
logging of errors out of loadModel, which is nice.