Add site XML validation to build.py
This will help ensure the schema is correct and in sync with the
source, as well as checking for errors in page authoring.

Right now, the schema is really constrained to exactly what we need
for now, including exactly the `xml:` attributes I need at the
moment. . If we start introducing HTML markup to our text elements,
however, we'll need to either extend the schema or relax it a bit with
element/attribute wildcards.

Because the validation currently takes several seconds, I offer an
command-line option to skip this step. For troubleshooting, say,
Javascript, skipping validation speeds things up dramatically.