Add support for Unicode transliteration
To do this, I have to process the site.xml yet again, finding and
converting all <al encoding="mdc"> elements.

I'm using the Classical Language Toolkit (CLTK) to do the
transformation, with a couple of tweaks explained in the script.
This means we now have Python library dependencies in our build
script, which I recommend creating a Python virtualenv to deal
with.  requirements.txt captures the new dependency.