Change model mousedown to click
This will get rid of most spurious (de)selection caused by camera
repositioning, as only a click-drag that starts and ends in
approximately the same place counts as a "click" event.

We can still do better, but only by really tracking the complete mouse
activity and (de)selecting only when a certain amount of drag is
detected.