Add location/source annotations to texts
Introducing a new element for this, text/desc. I'm
making it required for now, but it's not reflected
in HTML yet.

This ensures we've copied over all of the content from
iwefaa-complete-text.gly.

Note that, in cases where a location appeared to apply
across multiple texts (and not just multiple fragments),
where the GLY file just lists the location once, I copied
out the location for each text separately.

I also took a crack at a description for the Amenirdis text.