Validate XML well-formedness
This is a tricky because we need to do it before our XInclude
preprocessing, but we also need to take XIncludes into account so we
can find all the source files to validate!

XSLT to the rescue again: spit out a text file with just the XInclude
hrefs, then stomp through it to find all the additional files to
validate. Do this recursively in case we ever get fancier.