Use Java tool to accelerate XSLT processing
The previous approach invokes the Saxon CLI repeatedly for site.xml
and each page. Because each invocation of Saxon requires booting up a
JVM, this is relatively slow.

Saxon's recommendation for performance is to write your own Java CLI
that calls the Saxon code multiple times in one CLI run, so that the
JVM only has to be booted up once. This is indeed a lot faster.

We didn't have this problem before because we were using XIncludes and
the xsl:result-document feature to assemble everything into one input
and generate all the output we need in one Saxon CLI invocation. In a
sense, we've come full circle by once again needing to find a way to
do it all in one shot. Here, though, I think the complexity may be in
a more desirable place: the XSL is now much more straightforward,
which should make HTML template authoring easier, and the Java code
will only need to be updated if we need to alter the transformations
we need to do.